Updated room.hpp method documentation
Added a few notes to some methods documentation to provide a helpful explanation as to why certain parameters are passed as certain types. Most of the assumption is based on passing strings rather than ints or enums, where these conversions will be taken care of in the room class. Except for update_room_status(). This assumes an enum will be passed.